Changed ratings to double in backend

+++ src/java/dochunt/models/Review.java
@@ -14,7 +14,7 @@ public class Review {
     public String doctorAlias;
     public String doctorFirstName;
     public String doctorLastName;
-    public int rating;
+    public double rating;
     public String comments;
     public long date; // Stored as ms since epoch
 }
